Refactor app.js for separation of concerns
Refactor `app.js` to improve separation of concerns by moving MQTT and Socket.IO handling to separate files and updating dependencies.

* **MQTT Handling**
  - Move MQTT connection, event handlers, and subscription logic from `app.js` to `mqttClient.js`.
  - Create `mqttClient.js` to handle all MQTT-related logic.
  - Export necessary functions and variables for use in `app.js`.

* **Socket.IO Handling**
  - Move Socket.IO connection, event handlers, and emission logic from `app.js` to `socketClient.js`.
  - Create `socketClient.js` to handle all Socket.IO-related logic.
  - Export necessary functions and variables for use in `app.js`.

* **Dependencies Update**
  - Update versions of various dependencies in `package-lock.json` to their latest versions.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/harperreed/office-avatar-vrm?shareId=XXXX-XXXX-XXXX-XXXX).
